+++ tests/test_metadata.py
@@ -2,7 +2,6 @@ import os
 import unittest
 
 from vizier.datastore.metadata import DatasetMetadata
-from vizier.datastore.metadata import add_annotation, get_first, update_annotations
 
 METADATA_FILE = './data/metadata.yaml'
 
@@ -22,88 +21,88 @@ class TestDatasetMetadata(unittest.TestCase):
     def test_copy_metadata(self):
         """Test adding and retrieving metadata for different object types."""
         ds_meta = DatasetMetadata()
-        add_annotation(ds_meta.for_column(1), 'comment', 'Some Comment')
-        add_annotation(ds_meta.for_row(0), 'name', 'Nonsense')
-        add_annotation(ds_meta.for_cell(1, 1), 'title', 'Nonsense')
+        ds_meta.for_column(1).add('comment', 'Some Comment')
+        ds_meta.for_row(0).add('name', 'Nonsense')
+        ds_meta.for_cell(1, 1).add('title', 'Nonsense')
         meta = ds_meta.copy_metadata()
         # Column annotation
         col_anno = meta.for_column(0)
-        self.assertEquals(len(col_anno), 0)
+        self.assertEquals(col_anno.size(), 0)
         col_anno = meta.for_column(1)
-        self.assertEquals(len(col_anno), 1)
-        self.assertEquals(get_first(col_anno, 'comment').value, 'Some Comment')
+        self.assertEquals(col_anno.size(), 1)
+        self.assertEquals(col_anno.find_one('comment').value, 'Some Comment')
         # Row annotation
         row_anno = meta.for_row(0)
-        self.assertEquals(len(col_anno), 1)
-        self.assertEquals(get_first(row_anno, 'name').value, 'Nonsense')
+        self.assertEquals(col_anno.size(), 1)
+        self.assertEquals(row_anno.find_one('name').value, 'Nonsense')
         row_anno = meta.for_row(1)
-        self.assertEquals(len(row_anno), 0)
+        self.assertEquals(row_anno.size(), 0)
         # Cell Annotations
         cell_anno = meta.for_cell(1, 1)
-        self.assertEquals(len(cell_anno), 1)
-        self.assertEquals(get_first(cell_anno, 'title').value, 'Nonsense')
+        self.assertEquals(cell_anno.size(), 1)
+        self.assertEquals(cell_anno.find_one('title').value, 'Nonsense')
         # Ensure that changes to the copy don't affect the original
-        add_annotation(meta.for_column(1), 'comment', 'New Comment')
-        add_annotation(meta.for_column(2), 'comment', 'Some Comment')
-        self.assertEquals(len(meta.for_column(1)), 2)
-        self.assertEquals(len(ds_meta.for_column(1)), 1)
-        self.assertEquals(get_first(ds_meta.for_column(1), 'comment').value, 'Some Comment')
+        meta.for_column(1).add('comment', 'New Comment')
+        meta.for_column(2).add('comment', 'Some Comment')
+        self.assertEquals(meta.for_column(1).size(), 2)
+        self.assertEquals(ds_meta.for_column(1).size(), 1)
+        self.assertEquals(ds_meta.for_column(1).find_one('comment').value, 'Some Comment')
         values = [a.value for a in meta.for_column(1).values()]
         self.assertTrue('Some Comment' in values)
         self.assertTrue('New Comment' in values)
-        self.assertEquals(len(ds_meta.for_column(2)), 0)
-        self.assertEquals(len(meta.for_column(2)), 1)
-        self.assertEquals(get_first(meta.for_column(2), 'comment').value, 'Some Comment')
+        self.assertEquals(ds_meta.for_column(2).size(), 0)
+        self.assertEquals(meta.for_column(2).size(), 1)
+        self.assertEquals(meta.for_column(2).find_one('comment').value, 'Some Comment')
 
     def test_dataset_metadata(self):
         """Test adding and retrieving metadata for different object types."""
         meta = DatasetMetadata()
-        add_annotation(meta.for_column(1), 'comment', 'Some Comment')
-        add_annotation(meta.for_row(0), 'name', 'Nonsense')
-        add_annotation(meta.for_cell(1, 1), 'title', 'Nonsense')
+        meta.for_column(1).add('comment', 'Some Comment')
+        meta.for_row(0).add('name', 'Nonsense')
+        meta.for_cell(1, 1).add('title', 'Nonsense')
         # Column annotation
         col_anno = meta.for_column(0)
-        self.assertEquals(len(col_anno), 0)
+        self.assertEquals(col_anno.size(), 0)
         col_anno = meta.for_column(1)
-        self.assertEquals(len(col_anno), 1)
-        self.assertEquals(get_first(col_anno, 'comment').value, 'Some Comment')
+        self.assertEquals(col_anno.size(), 1)
+        self.assertEquals(col_anno.find_one('comment').value, 'Some Comment')
         # Row annotation
         row_anno = meta.for_row(0)
-        self.assertEquals(len(col_anno), 1)
-        self.assertEquals(get_first(row_anno, 'name').value, 'Nonsense')
+        self.assertEquals(col_anno.size(), 1)
+        self.assertEquals(row_anno.find_one('name').value, 'Nonsense')
         row_anno = meta.for_row(1)
-        self.assertEquals(len(row_anno), 0)
+        self.assertEquals(row_anno.size(), 0)
         # Cell Annotations
         cell_anno = meta.for_cell(1, 1)
-        self.assertEquals(len(cell_anno), 1)
-        self.assertEquals(get_first(cell_anno, 'title').value, 'Nonsense')
+        self.assertEquals(cell_anno.size(), 1)
+        self.assertEquals(cell_anno.find_one('title').value, 'Nonsense')
 
     def test_io_metadata(self):
         """Test reading and writing metadata from/to file."""
         ds_meta = DatasetMetadata()
-        add_annotation(ds_meta.for_column(1), 'comment', 'Some Comment')
-        add_annotation(ds_meta.for_column(1), 'name', 'foo')
-        add_annotation(ds_meta.for_row(0), 'name', 'Nonsense')
-        add_annotation(ds_meta.for_cell(1, 1), 'title', 'Nonsense')
+        ds_meta.for_column(1).add('comment', 'Some Comment')
+        ds_meta.for_column(1).add('name', 'foo')
+        ds_meta.for_row(0).add('name', 'Nonsense')
+        ds_meta.for_cell(1, 1).add('title', 'Nonsense')
         ds_meta.to_file(METADATA_FILE)
         meta = DatasetMetadata.from_file(METADATA_FILE)
         # Column annotation
         col_anno = meta.for_column(0)
-        self.assertEquals(len(col_anno), 0)
+        self.assertEquals(col_anno.size(), 0)
         col_anno = meta.for_column(1)
-        self.assertEquals(len(col_anno), 2)
-        self.assertEquals(get_first(col_anno, 'comment').value, 'Some Comment')
-        self.assertEquals(get_first(col_anno, 'name').value, 'foo')
+        self.assertEquals(col_anno.size(), 2)
+        self.assertEquals(col_anno.find_one('comment').value, 'Some Comment')
+        self.assertEquals(col_anno.find_one('name').value, 'foo')
         # Row annotation
         row_anno = meta.for_row(0)
-        self.assertEquals(len(row_anno), 1)
-        self.assertEquals(get_first(row_anno, 'name').value, 'Nonsense')
+        self.assertEquals(row_anno.size(), 1)
+        self.assertEquals(row_anno.find_one('name').value, 'Nonsense')
         row_anno = meta.for_row(1)
-        self.assertEquals(len(row_anno), 0)
+        self.assertEquals(row_anno.size(), 0)
         # Cell Annotations
         cell_anno = meta.for_cell(1, 1)
-        self.assertEquals(len(cell_anno), 1)
-        self.assertEquals(get_first(cell_anno, 'title').value, 'Nonsense')
+        self.assertEquals(cell_anno.size(), 1)
+        self.assertEquals(cell_anno.find_one('title').value, 'Nonsense')
         annotated_cells = ds_meta.cells_with_annotations()
         self.assertEquals(len(annotated_cells), 1)
         self.assertEquals(annotated_cells[0]['column'], 1)
@@ -113,27 +112,27 @@ class TestDatasetMetadata(unittest.TestCase):
         """Test update annotation statements."""
         meta = DatasetMetadata()
         # Create a new annotation
-        anno1 = update_annotations(meta.for_column(1), key='comment', value='Some Comment')
-        self.assertEquals(get_first(meta.for_column(1), 'comment').identifier, anno1.identifier)
-        self.assertEquals(get_first(meta.for_column(1), 'comment').key, 'comment')
-        self.assertEquals(get_first(meta.for_column(1), 'comment').value, 'Some Comment')
+        anno1 = meta.for_column(1).update(key='comment', value='Some Comment')
+        self.assertEquals(meta.for_column(1).find_one('comment').identifier, anno1.identifier)
+        self.assertEquals(meta.for_column(1).find_one('comment').key, 'comment')
+        self.assertEquals(meta.for_column(1).find_one('comment').value, 'Some Comment')
         # Create another annotation
-        anno2 = update_annotations(meta.for_column(1), key='comment', value='Some Comment')
-        self.assertEquals(len(meta.for_column(1)), 2)
+        anno2 = meta.for_column(1).update(key='comment', value='Some Comment')
+        self.assertEquals(meta.for_column(1).size(), 2)
         # Update first annotation value
-        anno1_2 = update_annotations(meta.for_column(1), identifier=anno1.identifier, value='Other Comment')
+        anno1_2 = meta.for_column(1).update(identifier=anno1.identifier, value='Other Comment')
         self.assertEquals(anno1_2.value, 'Other Comment')
-        self.assertEquals(meta.for_column(1)[anno1.identifier].value, 'Other Comment')
-        self.assertEquals(meta.for_column(1)[anno2.identifier].value, 'Some Comment')
+        self.assertEquals(meta.for_column(1).get(anno1.identifier).value, 'Other Comment')
+        self.assertEquals(meta.for_column(1).get(anno2.identifier).value, 'Some Comment')
         # Delete the second annotation
-        update_annotations(meta.for_column(1), identifier=anno2.identifier)
-        self.assertEquals(len(meta.for_column(1)), 1)
-        self.assertEquals(get_first(meta.for_column(1), 'comment').value, 'Other Comment')
+        meta.for_column(1).update(identifier=anno2.identifier)
+        self.assertEquals(meta.for_column(1).size(), 1)
+        self.assertEquals(meta.for_column(1).find_one('comment').value, 'Other Comment')
         # Update the key of the remaining annotation
-        update_annotations(meta.for_column(1), identifier=anno1.identifier, key='remark', value='Other Comment')
-        self.assertEquals(len(meta.for_column(1)), 1)
-        self.assertIsNone(get_first(meta.for_column(1), 'commanr'))
-        self.assertEquals(get_first(meta.for_column(1), 'remark').value, 'Other Comment')
+        meta.for_column(1).update(identifier=anno1.identifier, key='remark', value='Other Comment')
+        self.assertEquals(meta.for_column(1).size(), 1)
+        self.assertIsNone(meta.for_column(1).find_one('commanr'))
+        self.assertEquals(meta.for_column(1).find_one('remark').value, 'Other Comment')
 
 
 if __name__ == '__main__':